+++ source/Node/CL2DFlex/pipeline/CL2DPipelineLetterbox.cpp
@@ -72,7 +72,7 @@ QCStatus_e CL2DPipelineLetterbox::Execute( ImageDescriptor_t &input, ImageDescri
         }
         else
         {
-            uint32_t srcOffset = input.offset;
+            uint32_t srcOffset = (uint32_t) input.offset;
             uint32_t sizeOne = (uint32_t) ( output.size ) / ( output.batchSize );
             uint32_t dstOffset = (uint32_t) ( output.offset ) + m_inputId * sizeOne;
