+++ source/Node/CL2DFlex/pipeline/CL2DPipelineResize.cpp
@@ -101,7 +101,7 @@ QCStatus_e CL2DPipelineResize::Execute( ImageDescriptor_t &input, ImageDescripto
         }
         else
         {
-            uint32_t srcOffset = input.offset;
+            uint32_t srcOffset = (uint32_t) input.offset;
             uint32_t sizeOne = (uint32_t) ( output.size ) / ( output.batchSize );
             uint32_t dstOffset = (uint32_t) ( output.offset ) + m_inputId * sizeOne;
